pyproject.toml
[project]
name = "deep-learning-template"
version = "0.1.0"
description = "template project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchvision==0.21.0",
    "torchcodec>=0.8.1",
    "transformers>=4.40.0",
    "tokenizers>=0.19.0",
    "huggingface-hub>=0.34,<1.0",
    "librosa>=0.11.0",
    "praat-parselmouth>=0.4.6",
    "matplotlib>=3.0.0",
    "seaborn>=0.13.2",
    "pandas>=2.3.3",
    "scipy>=1.16.3",
    "scikit-learn>=1.7.2",
    "nvitop>=1.6.0",
    "openpyxl>=3.1.5",
    "pyyaml>=6.0.2",
    "ipdb>=0.13.13",
    "opensmile>=2.6.0",
    "speechbrain>=1.0.3",
]

[project.urls]
Homepage = "https://github.com/Hex4C59/deep_learning_template"
Repository = "https://github.com/Hex4C59/deep_learning_template.git"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
